pyproject.toml
[project]
name = "azure-things"
version = "0.1.0"
description = "Things app-like interface for managing Azure Boards work items"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flet>=0.24.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
azure-things = "azure_things.main:main"

[tool.uv]
package = true
dev-dependencies = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
]
